Write a program that finds and prints the first 98 Fibonacci numbers, starting with 1 and 2, followed by a new line

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -6,12 +6,12 @@
  */
 void print_fibonacci(int n)
 {
-int current = 1, previous = 0, temp;
+unsigned long int current = 1, previous = 0, temp;
 int count = 0;
 
 while (count < n)
 {
-printf("%d", current);
+printf("%lu", current);
 count++;
 
 if (count < n)
